added mn tm test
